# docs/swagger.yaml
definitions:
  handlers.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      login:
        example: john_doe
        type: string
      password:
        example: securepassword123
        type: string
    required:
    - login
    - password
    type: object
...
  handlers.RegisterRequest:
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      login:
        example: john_doe
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: securepassword123
        type: string
    required:
    - login
    - name
    - password
    type: object
...
  handlers.UpdateProfileRequest:
    properties:
      avatarUrl:
        example: https://example.com/new-avatar.jpg
        type: string
      name:
        example: John Smith
        type: string
    type: object
  handlers.UserResponse:
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      lastSeen:
        example: "2024-01-15T10:30:00Z"
        type: string
      login:
        example: john_doe
        type: string
      name:
        example: John Doe
        type: string
      online:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Вход в систему с логином и паролем
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
  /user/me:
    get:
      description: Возвращает информацию о текущем авторизованном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить профиль текущего пользователя
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Обновляет информацию профиля текущего пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль пользователя
      tags:
      - user
swagger: "2.0"